pyproject.toml
[project]
name = "oztech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.23",
    "django-cors-headers>=4.7.0",
    "django-unfold>=0.63.0",
    "djangorestframework>=3.16.1",
    "firebase-admin>=7.1.0",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "requests>=2.32.4",
]
